+++ src/onegov.swissvotes/onegov/swissvotes/models/policy_area.py
@@ -388,9 +388,8 @@ class PolicyArea(object):
             lookup = lookup.get('children', {})
         return result
 
-    def html(self, request, full=False):
-        label_path = [request.translate(part) for part in self.label_path]
-        title = ' &gt; '.join(label_path)
-        return f'<span>{title}' if full else (
-            f'<span title="{title}">{label_path[-1]}</span>'
-        )
+    def html(self, request):
+        title = ' &gt; '.join([
+            request.translate(part) for part in self.label_path
+        ])
+        return f'<span>{title}</span>'